car_price_project/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
joblib==1.4.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0
